requirements.txt
fastapi
uvicorn
requests
httpx
python-dotenv
twilio
openai
python-multipart
